final version

+++ app/components/Card/Card.tsx
@@ -1,28 +1,27 @@
 'use client'
-import React, { useEffect, useState } from 'react';
-import Link from 'next/link';
-import { useAppDispatch, useAppSelector } from '@/redux/hooks';
-import { Property } from '@/redux/features/SelecSlice';
 
-import { Post, useAddFavoriteMutation, useDeleteFavoriteMutation} from '@/redux/services/favorite';
+
+import { useState } from 'react';
+import Link from 'next/link';
+import { useAppDispatch,useAppSelector } from '@/redux/hooks';
+import { Property } from '@/redux/features/SelecSlice';
+import { Post, useAddFavoriteMutation, useDeleteFavoriteMutation, useGetFavoritesQuery } from '@/redux/services/favorite';
 import StarRating from '../StarRating/StarRating';
 import { getFavorite } from '@/redux/features/Favorite';
-
 
 
 interface CardsProps {  
   properties: Property;
 }
 
-const Card: React.FC<CardsProps> = ({properties}) => {
+const Card: React.FC<CardsProps> = ({ properties }) => {
   const [currentImage, setCurrentImage] = useState(0);
   const [isFavorite, setIsFavorite] = useState(false);
   const dispatch = useAppDispatch();
-  const user = useAppSelector((state) => state.user.user);
-
+  
   const [deleteFavorite]=useDeleteFavoriteMutation()
   const [addFavorite]=useAddFavoriteMutation()
-  
+  const user = useAppSelector((state) => state.user.user);
 
   const nextImage = () => {
     if (currentImage < properties.images.length - 1) {
@@ -35,45 +34,66 @@ const Card: React.FC<CardsProps> = ({properties}) => {
       setCurrentImage(currentImage - 1);
     }
   };
-
-
-
-  const { id, title, price, images } = properties;
-  const userId = "e28a65e9-82e6-4dc9-8997-ddcfdc671c7f";
-  const postId=id
+  
+  const { id, title, price, images } = properties;  
+  
+  const userId = user ? user?.id : null
+  const postId = id
 
   const toggleFavorite = async() => {
+    if(!userId) return;
+
     setIsFavorite(!isFavorite);
     
     if (isFavorite) {
       deleteFavorite({ userId, postId })
-      // dispatch(getFavorite(userId));
-    } else {
-       
-              const post: Post = {
-                userId,
-                postId,
-                images,
-                title,
-                price,
-              };
-            
-              addFavorite(post);
-              
- 
+    } else { 
+      const post: Post = {
+        userId,
+        postId,
+        images,
+        title,
+        price,
+      };
+      addFavorite(post);
     }
+  } else {
+    const post: Post = {
+      userId: userId || "",
+      postId,
+      images,
+      title,
+      price,
+    };
+    addFavorite(post);
+  }
 
-               await dispatch(getFavorite(userId));
-  };
-      
+  // Asegurarte de que userId no sea undefined antes de llamar a dispatch
+  if (userId) {
+    await dispatch(getFavorite(userId));
+  };  
+//     } else {
+       
+//               const post: Post = {
+//                 userId,
+//                 postId,
+//                 images,
+//                 title,
+//                 price,
+//               };            
+//               addFavorite(post);        
+ 
+//     }
+//                await dispatch(getFavorite(userId));
+   };     
+
 
   const favoriteImageUrl = '/dislike.png';
   const notFavoriteImageUrl = '/like.png';
-  console.log("user",user);
-  
+  console.log("user",user); 
 
   return (
-    <div className="w-96 p-4 bg-white rounded-3xl shadow-md transform hover:scale-105 transition-transform duration-300 ease-in-out">
+    <div className="w-80 sm:w-96 p-4 bg-white rounded-3xl shadow-md transform hover:scale-105 transition-transform duration-300 ease-in-out">
       <div className="h-52 w-90 relative">
         {properties.images?.map((imagen, index) => (
           <img
@@ -115,20 +135,21 @@ const Card: React.FC<CardsProps> = ({properties}) => {
       <div className='flex flex-col justify-around gap-8'>
         <div>
           <h2 className="text-xl font-bold text-center">
-            <Link href={`/Views/${properties.id}`}>{properties.title}</Link>
+            {/* <Link href={`/Views/${properties.id}`}> */}<p>{ properties.title }</p>{/* </Link> */}
           </h2>
-          <h2 className='text-center mt-5 text-xl font-semibold'>${properties.price}</h2>
+          <h2 className='text-center mt-5 text-xl font-semibold'>${ properties.price }</h2>
         </div>
         <div className="p-4">
           {/* Utiliza el componente StarRating para mostrar el puntaje como estrellas */}
-          <StarRating score={properties.score} />
-          <p className="text-gray-600">{properties.condition}</p>
-          <p className="text-gray-600">
-            <img src="/location .png" width={30} height={30} alt="direccion" /> {properties.type} en {properties.streetName} {properties.floorNumber}
+          <StarRating score={ properties.score ? properties.score : 0 } />
+          {/* <p className="text-gray-600">{properties.condition}</p> */}
+          <p className="text-gray-600 flex items-center">
+            Dirección: { properties.streetName } { properties.floorNumber }
           </p>
-          <p className="text-gray-600">{properties.country}, {properties.city}</p>
-          <div className="flex justify-end items-center mt-4">
-            <button onClick={toggleFavorite} className={`favorite-button ${isFavorite ? 'favorite' : ''}`}>
+          <p className="text-gray-600">Ubicación: { properties.country }, { properties.city }</p>
+          <div className="flex justify-between items-center mt-5">
+            <Link href={ `/Views/${properties.id}` }><button className="text-white bg-[#FD8974] hover:bg-[#E07564] font-medium rounded-lg text-sm px-5 py-2.5 text-center rounded-full">Mas Detalle</button></Link>
+            <button onClick={ toggleFavorite } className={ `favorite-button ${isFavorite ? 'favorite' : ''}` }>
               {isFavorite ? (
                 <img src={favoriteImageUrl} width={40} height={40} alt="Favorito" />
               ) : (
@@ -139,9 +160,8 @@ const Card: React.FC<CardsProps> = ({properties}) => {
         </div>
       </div>
     </div>
-  );
-};
+    );
+  };
+
 
 export default Card;
-
-
